Update phaser.plugin.isometric.d.ts
I expanded on @edgarjcfn original work. There are some issues as detailed at the top of the file which are not done mostly for the same reason as before (opened ended classes). I still hope to get time to play with this library in the next week or so to give it a test :D